requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
bcrypt==4.1.2
SQLAlchemy==2.0.30
gunicorn==20.1.0
flask-migrate==4.0.4
itsdangerous==2.1.2
flask-mail==0.9.1
Werkzeug==2.3.7
PyJWT==2.6.0
Flask-JWT-extended==4.6.0
apscheduler==3.10.1
marshmallow==3.21.1
email-validator==2.1.1